+++ client/Components/GameBoard/SquishableCardPanel.jsx
@@ -36,6 +36,7 @@ class SquishableCardPanel extends React.Component {
 
             return (<Card key={ card.uuid }
                 card={ card }
+                cardBackUrl={ this.props.cardBackUrl }
                 disableMouseOver={ !card.name }
                 canDrag={ this.props.manualMode }
                 onClick={ this.props.onCardClick }
@@ -114,6 +115,7 @@ class SquishableCardPanel extends React.Component {
 
 SquishableCardPanel.displayName = 'SquishableCardPanel';
 SquishableCardPanel.propTypes = {
+    cardBackUrl: PropTypes.string,
     cardSize: PropTypes.string,
     cards: PropTypes.array,
     className: PropTypes.string,
