+++ server/game/Card.js
@@ -56,7 +56,7 @@ class Card extends EffectSource {
         this.printedHouse = cardData.house;
         this.cardPrintedAmber = cardData.amber;
         this.maverick = cardData.maverick;
-        this.anomoly = cardData.anomoly;
+        this.anomaly = cardData.anomaly;
 
         this.upgrades = [];
         this.parent = null;
@@ -849,7 +849,7 @@ class Card extends EffectSource {
 
         // Include card specific information useful for UI rendering
         result.maverick = this.maverick;
-        result.anomoly = this.anomoly;
+        result.anomaly = this.anomaly;
         result.cardPrintedAmber = this.cardPrintedAmber;
         result.locale = this.locale;
         return result;
@@ -870,7 +870,7 @@ class Card extends EffectSource {
         }
 
         let state = {
-            anomoly: this.anomoly,
+            anomaly: this.anomaly,
             id: this.cardData.id,
             image: this.cardData.image,
             canPlay: (activePlayer === this.game.activePlayer) && this.game.activePlayer.activeHouse &&
